Replace 'only_new' with 'exclude'
The idea is that, rather than having one boolean choice, we can
define a number of string values, such that we can also (for instance)
only predict 'old' links (for e.g. anomaly detection).

If we're directly using a predictor (say, CommonNeighbours), this means one
should change a line like this:

    c = CommonNeighbours(G, only_new=True)

to:

    c = CommonNeighbours(G, excluded=G.edges())